requirements.txt
flask 
flask-cors 
replicate 
gunicorn 
python-dotenv
